исправление ошибки при перерисовке попапа
