+++ src/presenter/film-card.js
@@ -2,6 +2,7 @@ import FilmCardView from '../view/film-card';
 import FilmDetailsPopupView from '../view/film-details';
 import FilmDetailsGenreView from '../view/film-details-genre';
 import FilmDetailsCommentView from '../view/film-details-comment';
+import FilmDetailsNewCommentView from '../view/film-details-new-comment';
 
 import { render, replace, remove } from '../utils/render';
 
@@ -35,6 +36,20 @@ export default class FilmCard {
 
     this._filmCardComponent = new FilmCardView(film);
     this._filmDetailsComponent = new FilmDetailsPopupView(film);
+
+    const filmDetailsGenresList = this._filmDetailsComponent.getElement().querySelectorAll('td:nth-last-of-type(1)')[FILM_DETAILS_GENRES_ROW_NUMBER];
+    const filmDetailsCommentWrap = this._filmDetailsComponent.getElement().querySelector('.film-details__comments-wrap');
+    const filmDetailsCommentsList = this._filmDetailsComponent.getElement().querySelector('.film-details__comments-list');
+
+    for (let i =0; i < this._film.genres.length; i++) {
+      render(filmDetailsGenresList, new FilmDetailsGenreView(this._film.genres[i]));
+    }
+
+    for (let i = 0; i < this._film.comments.length; i++) {
+      render(filmDetailsCommentsList, new FilmDetailsCommentView(this._film.comments[i]));
+    }
+
+    render(filmDetailsCommentWrap, new FilmDetailsNewCommentView);
 
     this._filmCardComponent.setAddToWatchListClickHandler(this._handleAddToWatchListClick);
     this._filmCardComponent.setMarkAsWatchedClickHandler(this._handleMarkAsWatchedClick);
@@ -75,17 +90,6 @@ export default class FilmCard {
 
     this._bodyElement.appendChild(this._filmDetailsComponent.getElement());
     this._bodyElement.classList.add('hide-overflow');
-
-    const filmDetailsGenresList = document.querySelectorAll('td:nth-last-of-type(1)')[FILM_DETAILS_GENRES_ROW_NUMBER];
-    const filmDetailsCommentsList = this._filmDetailsComponent.getElement().querySelector('.film-details__comments-list');
-
-    for (let i =0; i < this._film.genres.length; i++) {
-      render(filmDetailsGenresList, new FilmDetailsGenreView(this._film.genres[i]));
-    }
-
-    for (let i = 0; i < this._film.comments.length; i++) {
-      render(filmDetailsCommentsList, new FilmDetailsCommentView(this._film.comments[i]));
-    }
 
     document.addEventListener('keydown', this._escKeyDownHandler);
   }
